Moved vertices generation from genesis to shaping
