+++ src/world/terraformation/genesis.rs
@@ -14,14 +14,15 @@ use bracket_noise::prelude::{FastNoise, FractalType, NoiseType};
 use futures_lite::future;
 
 use crate::world::{
-    math, mesh,
+    math,
     storage::{
-        chunk::{self, Chunk, ChunkKind, ChunkNeighborhood},
-        voxel::{self, FacesOcclusion, KindsDescs, VoxelFace, VoxelVertex},
+        chunk::{self, Chunk, ChunkKind},
+        voxel::{self, KindsDescs},
         VoxWorld,
     },
-    terraformation::ChunkFacesOcclusion,
 };
+
+use super::shaping;
 
 const CACHE_PATH: &str = "cache/chunks/";
 const CACHE_EXT: &str = "bin";
@@ -59,7 +60,7 @@ fn setup_resources(mut commands: Commands, asset_server: Res<AssetServer>) {
     let input_path = format!("{}/assets/voxels/kind.ron", env!("CARGO_MANIFEST_DIR"));
     let f = std::fs::File::open(&input_path).expect("Failed opening kind descriptions file");
     let descs: KindsDescs = ron::de::from_reader(f).unwrap();
-    
+
     let atlas = asset_server.load(&descs.atlas_path);
 
     commands.insert_resource(KindsDescsRes { descs, atlas });
@@ -527,153 +528,16 @@ fn recompute_chunks(
     let mut result = vec![];
 
     for local in locals {
-        update_neighborhood(world, local);
+        if shaping::recompute_chunk(world, &kinds_descs, local) {
+            result.push(local);
 
-        if let Some(chunk) = world.get_mut(local) {
-            let occlusion = faces_occlusion(&chunk.kinds);
-            if !occlusion.is_fully_occluded() {
-                let faces = mesh::merge_faces(&occlusion, &chunk);
-                chunk.vertices = generate_vertices(faces, &kinds_descs);
-
-                result.push(local);
-            }
-
+            // TODO: Maybe save all chunks at the end of the batch will be better?
             let path = local_path(local);
-            save_chunk(&path, chunk);
+            save_chunk(&path, world.get(local).unwrap());
         }
     }
 
     result
-}
-
-/**
-Generates vertices data from a given [`VoxelFace`] list.
-
-All generated indices will be relative to a triangle list.
-
-**Returns** a list of generated [`VoxelVertex`].
-*/
-fn generate_vertices(faces: Vec<VoxelFace>, kinds_descs: &KindsDescs) -> Vec<VoxelVertex> {
-    perf_fn_scope!();
-
-    let mut vertices = vec![];
-    let tile_texture_size = 1.0 / kinds_descs.count_tiles() as f32;
-
-    for face in faces {
-        let normal = face.side.normal();
-
-        let face_desc = kinds_descs.get_face_desc(&face);
-        let tile_coord_start = face_desc.offset.as_vec2() * tile_texture_size;
-
-        let faces_vertices = face
-            .vertices
-            .iter()
-            .enumerate()
-            .map(|(i, v)| {
-                let base_vertex_idx = mesh::VERTICES_INDICES[face.side as usize][i];
-                let base_vertex: Vec3 = mesh::VERTICES[base_vertex_idx].into();
-
-                base_vertex + v.as_vec3()
-            })
-            .collect::<Vec<_>>();
-
-        debug_assert!(
-            faces_vertices.len() == 4,
-            "Each face should have 4 vertices"
-        );
-
-        fn calc_tile_size(min: Vec3, max: Vec3) -> f32 {
-            (min.x - max.x).abs() + (min.y - max.y).abs() + (min.z - max.z).abs()
-        }
-
-        let x_tile = calc_tile_size(faces_vertices[0], faces_vertices[1]) * tile_texture_size;
-        let y_tile = calc_tile_size(faces_vertices[0], faces_vertices[3]) * tile_texture_size;
-
-        let tile_uv = [
-            (0.0, y_tile).into(),
-            (x_tile, y_tile).into(),
-            (x_tile, 0.0).into(),
-            (0.0, 0.0).into(),
-        ];
-
-        for (i, v) in faces_vertices.into_iter().enumerate() {
-            vertices.push(VoxelVertex {
-                position: v,
-                normal,
-                uv: tile_uv[i],
-                tile_coord_start,
-            });
-        }
-    }
-
-    debug_assert!(!vertices.is_empty());
-    vertices
-}
-
-/**
-Computes the faces occlusion data of the given [`ChunkKind`]
-
-**Returns** computed [`ChunkFacesOcclusion`]
-*/
-fn faces_occlusion(chunk: &ChunkKind) -> ChunkFacesOcclusion {
-    perf_fn_scope!();
-
-    let mut occlusion = ChunkFacesOcclusion::default();
-    for voxel in chunk::voxels() {
-        let mut voxel_faces = FacesOcclusion::default();
-
-        if chunk.get(voxel).is_empty() {
-            voxel_faces.set_all(true);
-        } else {
-            for side in voxel::SIDES {
-                let dir = side.dir();
-                let neighbor_pos = voxel + dir;
-
-                let neighbor_kind = if !chunk::is_within_bounds(neighbor_pos) {
-                    let (_, next_chunk_voxel) = chunk::overlap_voxel(neighbor_pos);
-
-                    match chunk.neighborhood.get(side, next_chunk_voxel) {
-                        Some(k) => k,
-                        None => continue,
-                    }
-                } else {
-                    chunk.get(neighbor_pos)
-                };
-
-                voxel_faces.set(side, !neighbor_kind.is_empty());
-            }
-        }
-
-        occlusion.set(voxel, voxel_faces);
-    }
-
-    occlusion
-}
-
-/**
-Updates the [`ChunkNeighborhood`] of a given chunk local.
-This function updates any neighborhood data needed by chunk.
-
-Currently it only updates kind neighborhood data, but in the future, it may update light and other relevant data.
-*/
-fn update_neighborhood(world: &mut VoxWorld, local: IVec3) {
-    if !world.exists(local) {
-        return;
-    }
-
-    let mut neighborhood = ChunkNeighborhood::default();
-    for side in voxel::SIDES {
-        let dir = side.dir();
-        let neighbor = local + dir;
-
-        if let Some(neighbor_chunk) = world.get(neighbor) {
-            neighborhood.set(side, &neighbor_chunk.kinds);
-        }
-    }
-
-    if let Some(chunk) = world.get_mut(local) {
-        chunk.kinds.neighborhood = neighborhood;
-    }
 }
 
 /**
@@ -1237,204 +1101,6 @@ mod tests {
                 ChunkCmd::Unload((1, 6, 1).into())
             ]
         );
-    }
-
-    #[test]
-    fn update_neighborhood() {
-        let mut world = VoxWorld::default();
-
-        let center = (1, 1, 1).into();
-        let mut chunk = Chunk::default();
-        chunk.kinds.set_all(10.into());
-        world.add(center, chunk);
-
-        for side in voxel::SIDES {
-            let dir = side.dir();
-            let pos = center + dir;
-            let mut chunk = Chunk::default();
-            chunk.kinds.set_all((side as u16).into());
-            world.add(pos, chunk);
-        }
-
-        super::update_neighborhood(&mut world, center);
-        let chunk = world.get(center).unwrap();
-
-        for side in voxel::SIDES {
-            match side {
-                voxel::Side::Right => {
-                    for a in 0..chunk::Y_AXIS_SIZE {
-                        for b in 0..chunk::Z_AXIS_SIZE {
-                            assert_eq!(
-                                chunk
-                                    .kinds
-                                    .neighborhood
-                                    .get(side, (0, a as i32, b as i32).into()),
-                                Some((side as u16).into())
-                            );
-                        }
-                    }
-                }
-                voxel::Side::Left => {
-                    for a in 0..chunk::Y_AXIS_SIZE {
-                        for b in 0..chunk::Z_AXIS_SIZE {
-                            assert_eq!(
-                                chunk
-                                    .kinds
-                                    .neighborhood
-                                    .get(side, (chunk::X_END as i32, a as i32, b as i32).into()),
-                                Some((side as u16).into())
-                            );
-                        }
-                    }
-                }
-                voxel::Side::Up => {
-                    for a in 0..chunk::X_AXIS_SIZE {
-                        for b in 0..chunk::Z_AXIS_SIZE {
-                            assert_eq!(
-                                chunk
-                                    .kinds
-                                    .neighborhood
-                                    .get(side, (a as i32, 0, b as i32).into()),
-                                Some((side as u16).into())
-                            );
-                        }
-                    }
-                }
-                voxel::Side::Down => {
-                    for a in 0..chunk::X_AXIS_SIZE {
-                        for b in 0..chunk::Z_AXIS_SIZE {
-                            assert_eq!(
-                                chunk
-                                    .kinds
-                                    .neighborhood
-                                    .get(side, (a as i32, chunk::Y_END as i32, b as i32).into()),
-                                Some((side as u16).into())
-                            );
-                        }
-                    }
-                }
-                voxel::Side::Front => {
-                    for a in 0..chunk::X_AXIS_SIZE {
-                        for b in 0..chunk::Y_AXIS_SIZE {
-                            assert_eq!(
-                                chunk
-                                    .kinds
-                                    .neighborhood
-                                    .get(side, (a as i32, b as i32, 0).into()),
-                                Some((side as u16).into())
-                            );
-                        }
-                    }
-                }
-                voxel::Side::Back => {
-                    for a in 0..chunk::X_AXIS_SIZE {
-                        for b in 0..chunk::Y_AXIS_SIZE {
-                            assert_eq!(
-                                chunk
-                                    .kinds
-                                    .neighborhood
-                                    .get(side, (a as i32, b as i32, chunk::Z_END as i32).into()),
-                                Some((side as u16).into())
-                            );
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn faces_occlusion_occlude_empty_chunk() {
-        // Arrange
-        let chunk = ChunkKind::default();
-
-        // Act
-        let occlusions = super::faces_occlusion(&chunk);
-
-        // Assert
-        assert!(
-            occlusions.iter().all(|a| a.is_fully_occluded()),
-            "A chunk full of empty-kind voxels should be fully occluded"
-        );
-    }
-
-    #[test]
-    fn faces_occlusion() {
-        // Arrange
-        let mut chunk = ChunkKind::default();
-
-        // Top-Bottom occlusion
-        chunk.set((1, 1, 1).into(), 1.into());
-        chunk.set((1, 2, 1).into(), 1.into());
-
-        // Full occluded voxel at (10, 10, 10)
-        chunk.set((10, 10, 10).into(), 1.into());
-        chunk.set((9, 10, 10).into(), 1.into());
-        chunk.set((11, 10, 10).into(), 1.into());
-        chunk.set((10, 9, 10).into(), 1.into());
-        chunk.set((10, 11, 10).into(), 1.into());
-        chunk.set((10, 10, 9).into(), 1.into());
-        chunk.set((10, 10, 11).into(), 1.into());
-
-        // Act
-        let faces_occlusion = super::faces_occlusion(&chunk);
-
-        // Assert
-        let faces = faces_occlusion.get((1, 2, 1).into());
-
-        assert_eq!(
-            faces,
-            [false, false, false, true, false, false].into(),
-            "Only down face should be occluded by the bottom voxel"
-        );
-
-        let faces = faces_occlusion.get((1, 1, 1).into());
-
-        assert_eq!(
-            faces,
-            [false, false, true, false, false, false].into(),
-            "Only down face should be occluded by the bottom voxel"
-        );
-
-        let faces = faces_occlusion.get((10, 10, 10).into());
-
-        assert_eq!(
-            faces,
-            [true; voxel::SIDE_COUNT].into(),
-            "Voxel fully surrounded by another non-empty voxels should be fully occluded"
-        );
-    }
-
-    #[test]
-    fn faces_occlusion_neighborhood() {
-        let mut world = VoxWorld::default();
-
-        let mut top = Chunk::default();
-        top.kinds.set_all(2.into());
-
-        let mut down = Chunk::default();
-        down.kinds.set_all(3.into());
-
-        let mut center = Chunk::default();
-        center
-            .kinds
-            .set((0, chunk::Y_END as i32, 0).into(), 1.into());
-        center.kinds.set((1, 0, 1).into(), 1.into());
-
-        world.add((0, 1, 0).into(), top);
-        world.add((0, 0, 0).into(), center);
-        world.add((0, -1, 0).into(), down);
-
-        super::update_neighborhood(&mut world, (0, 0, 0).into());
-        let center = world.get((0, 0, 0).into()).unwrap();
-
-        let faces_occlusion = super::faces_occlusion(&center.kinds);
-
-        let faces = faces_occlusion.get((0, chunk::Y_END as i32, 0).into());
-        assert_eq!(faces, [false, false, true, false, false, false].into());
-
-        let faces = faces_occlusion.get((1, 0, 1).into());
-        assert_eq!(faces, [false, false, false, true, false, false].into());
     }
 
     // #[test]
